requirements.txt
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
cffi==1.14.5
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
Django==3.1.7
hyperlink==21.0.0
idna==3.1
image==1.5.33
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.2
parsel==1.6.0
Pillow==8.1.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pytz==2021.1
queuelib==1.5.0
redis==3.5.3
Scrapy==2.4.1
scrapy-redis==0.6.8
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
splinter==0.14.0
sqlparse==0.4.1
Twisted==20.3.0
urllib3==1.26.3
w3lib==1.22.0
zope.interface==5.2.0
